requirements.txt
pywhatkit==5.3
pandas==1.5.3
python-dotenv==0.21.0
openpyxl==3.0.10
